Refactored backend exceptions
