+++ backend/src/service/exceptions.py
@@ -1,0 +1,8 @@
+"""Custom exception classes"""
+
+
+class ValidationException(Exception):
+    """
+    This class is used to handle exceptions raised due to input validaiton.
+    It behaves identially to the Exception class.
+    """
